refactor: MVC 구조 변경
